+++ examples/shop_app_example/lib/main.dart
@@ -12,6 +12,7 @@ import 'package:talker_flutter/talker_flutter.dart';
 import 'package:talker_shop_app_example/repositories/products/products.dart';
 import 'package:talker_shop_app_example/ui/presentation_frame.dart';
 import 'package:talker_shop_app_example/ui/ui.dart';
+import 'package:talker_shop_app_example/utils/good_log.dart';
 import 'package:talker_shop_app_example/utils/scroll_behavior.dart';
 import 'package:talker_shop_app_example/utils/utils.dart';
 
@@ -32,7 +33,8 @@ void _initDependencies() {
   final talker = TalkerFlutter.init(
     settings: TalkerSettings(
       colors: {
-        TalkerLogType.verbose: AnsiPen()..yellow(),
+        TalkerLogType.verbose.key: AnsiPen()..yellow(),
+        GoodLog.getKey: GoodLog.getPen,
       },
     ),
   );
@@ -55,7 +57,7 @@ void _initDependencies() {
   DI.registerSingleton<AbstractProductsRepository>(
     ProductsRepository(dio: dio),
   );
-  talker.info('Repositories initialization completed');
+  talker.logCustom(GoodLog('Repositories initialization completed'));
 
   Bloc.observer = TalkerBlocObserver(
     talker: talker,