+++ packages/talker/lib/src/settings.dart
@@ -1,54 +1,68 @@
 import 'package:talker/talker.dart';
 
-const _defaultTitles = {
+final _defaultTitles = {
   /// Base logs section
-  TalkerLogType.critical: 'critical',
-  TalkerLogType.warning: 'warning',
-  TalkerLogType.verbose: 'verbose',
-  TalkerLogType.info: 'info',
-  TalkerLogType.debug: 'debug',
-  TalkerLogType.error: 'error',
-  TalkerLogType.exception: 'exception',
+  TalkerLogType.critical.key: 'critical',
+  TalkerLogType.warning.key: 'warning',
+  TalkerLogType.verbose.key: 'verbose',
+  TalkerLogType.info.key: 'info',
+  TalkerLogType.debug.key: 'debug',
+  TalkerLogType.error.key: 'error',
+  TalkerLogType.exception.key: 'exception',
 
   /// Http section
-  TalkerLogType.httpError: 'http-error',
-  TalkerLogType.httpRequest: 'http-request',
-  TalkerLogType.httpResponse: 'http-response',
+  TalkerLogType.httpError.key: 'http-error',
+  TalkerLogType.httpRequest.key: 'http-request',
+  TalkerLogType.httpResponse.key: 'http-response',
 
   /// Bloc section
-  TalkerLogType.blocEvent: 'bloc-event',
-  TalkerLogType.blocTransition: 'bloc-transition',
-  TalkerLogType.blocCreate: 'bloc-create',
-  TalkerLogType.blocClose: 'bloc-close',
+  TalkerLogType.blocEvent.key: 'bloc-event',
+  TalkerLogType.blocTransition.key: 'bloc-transition',
+  TalkerLogType.blocCreate.key: 'bloc-create',
+  TalkerLogType.blocClose.key: 'bloc-close',
+
+  /// Riverpod section
+  TalkerLogType.riverpodAdd.key: 'riverpod-add',
+  TalkerLogType.riverpodUpdate.key: 'riverpod-update',
+  TalkerLogType.riverpodDispose.key: 'riverpod-dispose',
+  TalkerLogType.riverpodFail.key: 'riverpod-fail',
 
   /// Flutter section
-  TalkerLogType.route: 'route',
+  TalkerLogType.route.key: 'route',
 };
 
 final _defaultColors = {
   /// Base logs section
-  TalkerLogType.critical: AnsiPen()..red(),
-  TalkerLogType.warning: AnsiPen()..yellow(),
-  TalkerLogType.verbose: AnsiPen()..gray(),
-  TalkerLogType.info: AnsiPen()..blue(),
-  TalkerLogType.debug: AnsiPen()..gray(),
-  TalkerLogType.error: AnsiPen()..red(),
-  TalkerLogType.exception: AnsiPen()..red(),
+  TalkerLogType.critical.key: AnsiPen()..red(),
+  TalkerLogType.warning.key: AnsiPen()..yellow(),
+  TalkerLogType.verbose.key: AnsiPen()..gray(),
+  TalkerLogType.info.key: AnsiPen()..blue(),
+  TalkerLogType.debug.key: AnsiPen()..gray(),
+  TalkerLogType.error.key: AnsiPen()..red(),
+  TalkerLogType.exception.key: AnsiPen()..red(),
 
   /// Http section
-  TalkerLogType.httpError: AnsiPen()..red(),
-  TalkerLogType.httpRequest: AnsiPen()..xterm(219),
-  TalkerLogType.httpResponse: AnsiPen()..xterm(46),
+  TalkerLogType.httpError.key: AnsiPen()..red(),
+  TalkerLogType.httpRequest.key: AnsiPen()..xterm(219),
+  TalkerLogType.httpResponse.key: AnsiPen()..xterm(46),
 
   /// Bloc section
-  TalkerLogType.blocEvent: AnsiPen()..xterm(51),
-  TalkerLogType.blocTransition: AnsiPen()..xterm(49),
-  TalkerLogType.blocCreate: AnsiPen()..xterm(35),
-  TalkerLogType.blocClose: AnsiPen()..xterm(198),
+  TalkerLogType.blocEvent.key: AnsiPen()..xterm(51),
+  TalkerLogType.blocTransition.key: AnsiPen()..xterm(49),
+  TalkerLogType.blocCreate.key: AnsiPen()..xterm(35),
+  TalkerLogType.blocClose.key: AnsiPen()..xterm(198),
+
+  /// Riverpod section
+  TalkerLogType.riverpodAdd.key: AnsiPen()..xterm(51),
+  TalkerLogType.riverpodUpdate.key: AnsiPen()..xterm(49),
+  TalkerLogType.riverpodDispose.key: AnsiPen()..xterm(198),
+  TalkerLogType.riverpodFail.key: AnsiPen()..red(),
 
   /// Flutter section
-  TalkerLogType.route: AnsiPen()..xterm(135),
+  TalkerLogType.route.key: AnsiPen()..xterm(135),
 };
+
+final _fallbackPen = AnsiPen()..gray();
 
 /// {@template talker_settings}
 /// This class used for setup [Talker] configuration
@@ -59,15 +73,21 @@ class TalkerSettings {
     bool useHistory = true,
     bool useConsoleLogs = true,
     int maxHistoryItems = 1000,
-    this.titles = _defaultTitles,
-    Map<TalkerLogType, AnsiPen>? colors,
+    Map<String, String>? titles,
+    Map<String, AnsiPen>? colors,
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
   })  : _useHistory = useHistory,
         _useConsoleLogs = useConsoleLogs,
-        _maxHistoryItems = maxHistoryItems {
+        _maxHistoryItems = maxHistoryItems,
+        _timeFormat = timeFormat {
     if (colors != null) {
       _defaultColors.addAll(colors);
     }
+    if (titles != null) {
+      _defaultTitles.addAll(titles);
+    }
     this.colors.addAll(_defaultColors);
+    this.titles.addAll(_defaultTitles);
   }
   // _writeToFile = writeToFile;
 
@@ -79,13 +99,17 @@ class TalkerSettings {
   final bool _useHistory;
 
   /// By default talker print all Errors / Exceptions and logs in console.
-  /// If [true] - printing in history [false] - not printing.
+  /// If [true] - printing in console [false] - not printing.
   bool get useConsoleLogs => _useConsoleLogs && enabled;
   final bool _useConsoleLogs;
 
   /// Max records count in history list
   int get maxHistoryItems => _maxHistoryItems;
   final int _maxHistoryItems;
+
+  /// The time format of the logs [TimeFormat]
+  TimeFormat get timeFormat => _timeFormat;
+  final TimeFormat _timeFormat;
 
   /// Use writing talker records in file
   // bool get writeToFile => _writeToFile && enabled;
@@ -106,9 +130,9 @@ class TalkerSettings {
   ///
   /// ```dart
   /// final customTitles = {
-  ///   TalkerTitle.info: "Information",
-  ///   TalkerTitle.error: "Error",
-  ///   TalkerTitle.warning: "Warning",
+  ///   TalkerTitle.info.key: "Information",
+  ///   TalkerTitle.error.key: "Error",
+  ///   TalkerTitle.warning.key: "Warning",
   /// };
   ///
   /// final logger = Talker(
@@ -117,7 +141,7 @@ class TalkerSettings {
   ///   )
   /// );
   /// ```
-  final Map<TalkerLogType, String> titles;
+  final Map<String, String> titles = _defaultTitles;
 
   /// Custom Logger Colors.
   ///
@@ -129,9 +153,9 @@ class TalkerSettings {
   ///
   /// ```dart
   /// final customColors = {
-  ///   TalkerKey.info: AnsiPen()..white(bold: true),
-  ///   TalkerKey.error: AnsiPen()..red(bold: true),
-  ///   TalkerKey.warning: AnsiPen()..yellow(bold: true),
+  ///   TalkerKey.info.key: AnsiPen()..white(bold: true),
+  ///   TalkerKey.error.key: AnsiPen()..red(bold: true),
+  ///   TalkerKey.warning.key: AnsiPen()..yellow(bold: true),
   /// };
   ///
   /// final logger = Talker(
@@ -142,14 +166,17 @@ class TalkerSettings {
   /// ```
   ///
   /// By using the `colors` field, you can customize the text colors for specific log keys in the console.
-  final Map<TalkerLogType, AnsiPen> colors = _defaultColors;
+  final Map<String, AnsiPen> colors = _defaultColors;
 
-  String getTitleByLogType(TalkerLogType key) {
-    return titles[key] ?? key.key;
+  String getTitleByLogKey(String key) {
+    return titles[key] ?? key;
   }
 
-  AnsiPen getAnsiPenByLogType(TalkerLogType key, {TalkerData? logData}) {
-    return colors[key] ?? logData?.pen ?? (AnsiPen()..gray());
+  AnsiPen getAnsiPenByLogType(TalkerLogType type, {TalkerData? logData}) =>
+      getPenByLogKey(type.key, fallbackPen: logData?.pen);
+
+  AnsiPen getPenByLogKey(String key, {AnsiPen? fallbackPen}) {
+    return colors[key] ?? fallbackPen ?? _fallbackPen;
   }
 
   TalkerSettings copyWith({
@@ -157,8 +184,8 @@ class TalkerSettings {
     bool? useHistory,
     bool? useConsoleLogs,
     int? maxHistoryItems,
-    Map<TalkerLogType, String>? titles,
-    Map<TalkerLogType, AnsiPen>? colors,
+    Map<String, String>? titles,
+    Map<String, AnsiPen>? colors,
   }) {
     return TalkerSettings(
       useHistory: useHistory ?? _useHistory,