+++ packages/talker/lib/src/utils/utils.dart
@@ -1,2 +1,3 @@
 export 'error_handler.dart';
 export 'time_formatter.dart';
+export 'time_format.dart';