+++ packages/talker_bloc_logger/lib/bloc_logs.dart
@@ -20,13 +20,11 @@ class BlocEventLog extends TalkerLog {
   String get key => TalkerLogType.blocEvent.key;
 
   @override
-  String generateTextMessage() {
-    return _createMessage();
-  }
-
-  String _createMessage() {
+  String generateTextMessage({
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
+  }) {
     final sb = StringBuffer();
-    sb.write(displayTitleWithTime);
+    sb.write(displayTitleWithTime(timeFormat: timeFormat));
     sb.write('\n$message');
     return sb.toString();
   }
@@ -48,13 +46,11 @@ class BlocStateLog extends TalkerLog {
   String get key => TalkerLogType.blocTransition.key;
 
   @override
-  String generateTextMessage() {
-    return _createMessage();
-  }
-
-  String _createMessage() {
+  String generateTextMessage({
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
+  }) {
     final sb = StringBuffer();
-    sb.write(displayTitleWithTime);
+    sb.write(displayTitleWithTime(timeFormat: timeFormat));
     sb.write('\n$message');
     sb.write(
         '\n${'CURRENT state: ${settings.printStateFullData ? '\n${transition.currentState}' : transition.currentState.runtimeType}'}');
@@ -80,13 +76,11 @@ class BlocChangeLog extends TalkerLog {
   String get key => TalkerLogType.blocTransition.key;
 
   @override
-  String generateTextMessage() {
-    return _createMessage();
-  }
-
-  String _createMessage() {
+  String generateTextMessage({
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
+  }) {
     final sb = StringBuffer();
-    sb.write(displayTitleWithTime);
+    sb.write(displayTitleWithTime(timeFormat: timeFormat));
     sb.write('\n$message');
     sb.write(
         '\n${'CURRENT state: ${settings.printStateFullData ? '\n${change.currentState}' : change.currentState.runtimeType}'}');
@@ -108,13 +102,11 @@ class BlocCreateLog extends TalkerLog {
   String? get key => TalkerLogType.blocCreate.key;
 
   @override
-  String generateTextMessage() {
-    return _createMessage();
-  }
-
-  String _createMessage() {
+  String generateTextMessage({
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
+  }) {
     final sb = StringBuffer();
-    sb.write(displayTitleWithTime);
+    sb.write(displayTitleWithTime(timeFormat: timeFormat));
     sb.write('\n$message');
     return sb.toString();
   }
@@ -132,13 +124,11 @@ class BlocCloseLog extends TalkerLog {
   String? get key => TalkerLogType.blocClose.key;
 
   @override
-  String generateTextMessage() {
-    return _createMessage();
-  }
-
-  String _createMessage() {
+  String generateTextMessage({
+    TimeFormat timeFormat = TimeFormat.timeAndSeconds,
+  }) {
     final sb = StringBuffer();
-    sb.write(displayTitleWithTime);
+    sb.write(displayTitleWithTime(timeFormat: timeFormat));
     sb.write('\n$message');
     return sb.toString();
   }