+++ packages/talker_dio_logger/lib/talker_dio_logger_interceptor.dart
@@ -31,6 +31,9 @@ class TalkerDioLogger extends Interceptor {
     bool? printResponseData,
     bool? printResponseHeaders,
     bool? printResponseMessage,
+    bool? printErrorData,
+    bool? printErrorHeaders,
+    bool? printErrorMessage,
     bool? printRequestData,
     bool? printRequestHeaders,
     AnsiPen? requestPen,
@@ -41,6 +44,9 @@ class TalkerDioLogger extends Interceptor {
       printRequestData: printRequestData,
       printRequestHeaders: printRequestHeaders,
       printResponseData: printResponseData,
+      printErrorData: printErrorData,
+      printErrorHeaders: printErrorHeaders,
+      printErrorMessage: printErrorMessage,
       printResponseHeaders: printResponseHeaders,
       printResponseMessage: printResponseMessage,
       requestPen: requestPen,
@@ -55,6 +61,9 @@ class TalkerDioLogger extends Interceptor {
     RequestInterceptorHandler handler,
   ) {
     super.onRequest(options, handler);
+    if (!settings.enabled) {
+      return;
+    }
     final accepted = settings.requestFilter?.call(options) ?? true;
     if (!accepted) {
       return;
@@ -66,7 +75,7 @@ class TalkerDioLogger extends Interceptor {
         requestOptions: options,
         settings: settings,
       );
-      _talker.logTyped(httpLog);
+      _talker.logCustom(httpLog);
     } catch (_) {
       //pass
     }
@@ -75,6 +84,9 @@ class TalkerDioLogger extends Interceptor {
   @override
   void onResponse(Response response, ResponseInterceptorHandler handler) {
     super.onResponse(response, handler);
+    if (!settings.enabled) {
+      return;
+    }
     final accepted = settings.responseFilter?.call(response) ?? true;
     if (!accepted) {
       return;
@@ -86,7 +98,7 @@ class TalkerDioLogger extends Interceptor {
         settings: settings,
         response: response,
       );
-      _talker.logTyped(httpLog);
+      _talker.logCustom(httpLog);
     } catch (_) {
       //pass
     }
@@ -95,6 +107,13 @@ class TalkerDioLogger extends Interceptor {
   @override
   void onError(DioException err, ErrorInterceptorHandler handler) {
     super.onError(err, handler);
+    if (!settings.enabled) {
+      return;
+    }
+    final accepted = settings.errorFilter?.call(err) ?? true;
+    if (!accepted) {
+      return;
+    }
     try {
       final message = '${err.requestOptions.uri}';
       final httpErrorLog = DioErrorLog(
@@ -102,7 +121,7 @@ class TalkerDioLogger extends Interceptor {
         dioException: err,
         settings: settings,
       );
-      _talker.logTyped(httpErrorLog);
+      _talker.logCustom(httpErrorLog);
     } catch (_) {
       //pass
     }