+++ packages/talker_flutter/lib/src/extensions/extensions.dart
@@ -1,5 +1,3 @@
-export 'color.dart';
 export 'history.dart';
-export 'iterable.dart';
 export 'talker_data.dart';
 export 'talker_flutter.dart';