+++ packages/talker_flutter/lib/src/extensions/iterable.dart
@@ -1,8 +1,0 @@
-extension FirstWhereOrNullExtension<E> on Iterable<E> {
-  E? firstWhereOrNull(bool Function(E) test) {
-    for (E element in this) {
-      if (test(element)) return element;
-    }
-    return null;
-  }
-}