+++ packages/talker_flutter/lib/src/ui/talker_view.dart
@@ -165,7 +165,8 @@ class _TalkerViewState extends State<TalkerView> {
   }
 
   void _copyTalkerDataItemText(TalkerData data) {
-    final text = data.generateTextMessage();
+    final text =
+        data.generateTextMessage(timeFormat: widget.talker.settings.timeFormat);
     Clipboard.setData(ClipboardData(text: text));
     _showSnackBar(context, 'Log item is copied in clipboard');
   }
@@ -220,7 +221,7 @@ class _TalkerViewState extends State<TalkerView> {
 
   Future<void> _shareLogsInFile() async {
     await _controller.downloadLogsFile(
-      widget.talker.history.text,
+      widget.talker.history.text(timeFormat: widget.talker.settings.timeFormat),
     );
   }
 
@@ -234,7 +235,9 @@ class _TalkerViewState extends State<TalkerView> {
   }
 
   void _copyAllLogs(BuildContext context) {
-    Clipboard.setData(ClipboardData(text: widget.talker.history.text));
+    Clipboard.setData(ClipboardData(
+        text: widget.talker.history
+            .text(timeFormat: widget.talker.settings.timeFormat)));
     _showSnackBar(context, 'All logs copied in buffer');
   }
 }