+++ packages/talker_http_logger/lib/talker_http_logger.dart
@@ -17,7 +17,7 @@ class TalkerHttpLogger extends InterceptorContract {
     required BaseRequest request,
   }) async {
     final message = '${request.url}';
-    _talker.logTyped(HttpRequestLog(message, request: request));
+    _talker.logCustom(HttpRequestLog(message, request: request));
     return request;
   }
 
@@ -26,7 +26,13 @@ class TalkerHttpLogger extends InterceptorContract {
     required BaseResponse response,
   }) async {
     final message = '${response.request?.url}';
-    _talker.logTyped(HttpResponseLog(message, response: response));
+
+    if (response.statusCode >= 400 && response.statusCode < 600) {
+      _talker.logCustom(HttpErrorLog(message, response: response));
+    } else {
+      _talker.logCustom(HttpResponseLog(message, response: response));
+    }
+
     return response;
   }
 }
@@ -48,7 +54,7 @@ class HttpRequestLog extends TalkerLog {
   String get key => TalkerLogType.httpRequest.key;
 
   @override
-  String generateTextMessage() {
+  String generateTextMessage({TimeFormat timeFormat = TimeFormat.timeAndSeconds}) {
     var msg = '[$title] [${request.method}] $message';
 
     final headers = request.headers;
@@ -77,10 +83,44 @@ class HttpResponseLog extends TalkerLog {
   AnsiPen get pen => (AnsiPen()..xterm(46));
 
   @override
-  String get title => TalkerLogType.httpResponse.key;
+  String get key => TalkerLogType.httpResponse.key;
 
   @override
-  String generateTextMessage() {
+  String generateTextMessage({TimeFormat timeFormat = TimeFormat.timeAndSeconds}) {
+    var msg = '[$title] [${response.request?.method}] $message';
+
+    final headers = response.request?.headers;
+
+    msg += '\nStatus: ${response.statusCode}';
+
+    try {
+      if (headers?.isNotEmpty ?? false) {
+        final prettyHeaders = encoder.convert(headers);
+        msg += '\nHeaders: $prettyHeaders';
+      }
+    } catch (_) {
+      // TODO: add handling can`t convert
+    }
+    return msg;
+  }
+}
+
+class HttpErrorLog extends TalkerLog {
+  HttpErrorLog(
+    String title, {
+    required this.response,
+  }) : super(title);
+
+  final BaseResponse response;
+
+  @override
+  AnsiPen get pen => AnsiPen()..red();
+
+  @override
+  String get key => TalkerLogType.httpError.key;
+
+  @override
+  String generateTextMessage({TimeFormat timeFormat = TimeFormat.timeAndSeconds}) {
     var msg = '[$title] [${response.request?.method}] $message';
 
     final headers = response.request?.headers;