+++ utils/pub_get_all.sh
@@ -25,6 +25,14 @@ echo "talker_bloc_logger"
 cd ../talker_bloc_logger
 flutter pub get
 
+echo "talker_http_logger"
+cd ../talker_http_logger
+flutter pub get
+
+echo "talker_riverpod_logger"
+cd ../talker_riverpod_logger
+flutter pub get
+
 echo "shop_app_example"
 cd ../../examples/shop_app_example
 flutter pub get