+++ utils/publish_all.sh
@@ -19,3 +19,11 @@ cd ../../
 cd packages/talker_bloc_logger
 flutter pub publish
 cd ../../
+
+cd packages/talker_riverpod_logger
+flutter pub publish
+cd ../../
+
+cd packages/talker_http_logger
+flutter pub publish
+cd ../../